sepetate into converters file
